Fixed added space in series titles
This bug had nothing to do with Series, though that's where it was
primarily visible. Simply put, ripping the title from the markdown DOM
was causing excess spacing.

In places where this was used directly into
HTML *(such as in a template)*, it was fine and unnoticed. HTML simply
trimmed the excess whitespace.

In places where the text was combined, such as in series titles, the
whitespace became visible and ugly. This is of course, resolved.