Update index.md
Adding this line "define('FS_METHOD', 'direct');" in the wp-config.php file allows the user to install or upload images, themes, plugins, etc. The purpose of the change is that whenever the user tries to upload some file it asks for the FTP Credentials and when given it stops responding. Defining the "direct" method and setting the permission of the wp-config.php file to 777 solves the problem and the user is now allowed to upload to wordpress.